+++ release_notes.md
@@ -6,6 +6,8 @@
 * Fix a misusing of get_subdesc_accessor in group.
 * Fix a race condition that could result in a crash if a `LinkView` was
   destroyed while another thread was adjusting accessors on a `LinkListColumn`.
+* Fix crashes and general brokenness when constructing a Query, inserting a
+  column into the queried table, and then running the query.
 
 ### Breaking changes